requirements.txt
# requirements.txt for Telegram Submission Bot
python-telegram-bot[job-queue]~=21.0